separate out file_proxy_s3 into separate class files
